fuldstændig amazing kode, så meget generics og interfaces

+++ recipe/app/src/main/java/aau/sw8/model/Ingredient.java
@@ -4,19 +4,19 @@ package aau.sw8.model;
  * Created by jacob on 3/27/14.
  */
 public class Ingredient {
-    private int ingredientId;
+    private long ingredientId;
     private String singular;
     private String plural;
     private Category category;
 
-    public Ingredient(int ingredientId, String singular, String plural, Category category) {
+    public Ingredient(long ingredientId, String singular, String plural, Category category) {
         this.ingredientId = ingredientId;
         this.singular = singular;
         this.plural = plural;
         this.category = category;
     }
 
-    public int getId(){
+    public long getId(){
         return ingredientId;
     }
